requirements.txt
selenium==4.15.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
openpyxl==3.1.2
